requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Flask-WTF
Flask-Bcrypt
Flask-Login
python-dotenv
email-validator
feedparser